Fix: Deserialize rollingFiveHourLimit fields as Double to handle decimal values from API (#44)

+++ android/app/src/main/java/com/duchastel/simon/syntheticwidget/data/QuotaModels.kt
@@ -20,7 +20,7 @@ data class RollingFiveHourLimit(
     @SerialName("remaining")
     val remaining: Double,
     @SerialName("max")
-    val max: Int,
+    val max: Double,
     @SerialName("limited")
     val limited: Boolean,
 )
